Parse Downloads and Updates
If tests are running for the first time, (which they would likely be in a clean-room CI env), the parser will fail on updating crates.io + downloading deps.

Make sure we parse these lines.

In addition, the test result line has a filtered out clause, so parse that as well.